Move components from overview_screen package to presentation package
