+++ app/src/main/java/com/uxstate/presentation/components/FavoritePhoto.kt
@@ -1,4 +1,4 @@
-package com.uxstate.presentation.photo_screen.components
+package com.uxstate.presentation.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
